consumers/map/requirements.txt.txt
bokeh
geopandas
kafka-python